refactored to more OOP
