+++ duke.txt
@@ -1,0 +1,1 @@
+0. [D] [ ] eat (by mon)
