+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,25 @@
+package duke;
+import java.util.ArrayList;
+
+public class TaskList {
+
+    private ArrayList<Task> taskArray;
+
+    public TaskList() {
+        taskArray = new ArrayList<>();
+    }
+
+    public void addTask(Task task) {
+        taskArray.add(task);
+    }
+
+    public ArrayList<Task> getTaskArray() {
+        return taskArray;
+    }
+
+    public int getIndex() {
+        return taskArray.size();
+    }
+
+
+}
